Move `updateCustodyInfoInDB` from the blockchain service to the sync service and use the justified checkpoint (+ 1).
